change!: Use `gitoxide` `Repository` instead of `git2::Repository` (#16)
This comes with plenty of changes to the API of the
`last_seen_reference()` and to the lower-level methods that take
object ids (now `git::hash::ObjectId`.

Note that `git2` is still used internally for fetching and cloning.
This change was made to assure that at no time there are two open
repositories (once for git2, once for `gitoxide`), as this has the
potential to double resource usage for file handles.