fix: Consider all crates yanked if 'yanked = true'. (#19)
Previously, due to a missing test, a bug snuck in that would assume
that all lines in an added files must be new versions, marking them
as `Change::Added`. This ignored the fact that any line could also
carry a yanked crate, misrepresenting them.

This is now fixed, and yanked crates generally show up as
`Change::Yanked`.

One might take this as a hint that diffentiating by yank status might
not be that useful after all as it doesn't scale that well. Maybe
a future version changes how `Change` is represented.