fix: make fetches work again by using safe-performance mode of `git-repository`.
This fixes the 'zlib stream broken' issue when fetching crates.io
changes which was caused by `git-repository` configuring for
max-performance by default, which affects a crate used by `git2` as
well. For some reason, changing to `zlib-ng` as backend wasn't taken
kindly by `libgit2` causing it to fail after a short while of receiving
a pack from the remote.

The fix avoids making such modifications to the zlib crate, allowing
both crates, `git-repository` and `git2` to co-exist in the same
dependency tree.