fix: Ignore all changed files with an extension.
There are non-crate files that as far as we know all have file
extensions, as opposed to the crate files we are interested in, which do
not.

Thus skipping all files with extension helps us to get past the initial
commit which includes such files, like `.github/*.yml`.

Related to https://github.com/rust-lang/docs.rs/pull/1807#issuecomment-1235158502