fix: assure differences are handled exhaustively. (#26)
Previously it was possible to have multiple diffs in one crate
distributed over multiple commits to rightfully show up as multiple
hunks of modified and added lines only register the modified lines,
not the new ones (or the deleted ones for that matter).

This would cause updates or removals to be missed.

Now hunks of changes are exhaused properly, fixing this issue.