fix: correctly configure 'serde' features of dependencies (#33).
Due to surprising an to me still quite unclear behaviour `cargo`
will turn on the `serde` features in the `hex` and `smartstring`
dependencies when building it locally from a repository clone,
but won't do it when building the crates-io version of the crate
after adding it as dependency.

This fixes the issue by correctly configuring it the `serde`
feature.