Turns out that using the resource cache is too costly
It's a bit unclear why that is, but let's try to just read objects from
the ODB instead.

Unfortunately, even without going through a resource cache, we still have
50% performance loss.

Turns out the pack-cache has this much of an influence, and we changed
the variable names to `GIX_*` instead of `GITOXIDE_*`.